acueducto_backend/requirements.txt
blinker==1.9.0
blueprint==3.4.2
Brotli==1.1.0
cffi==1.17.1
chardet==5.2.0
click==8.1.7
colorama==0.4.6
cssselect2==0.7.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
fonttools==4.55.8
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
mysqlclient==2.2.6
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
PyMySQL==1.1.1
pyphen==0.17.2
python-dotenv==1.0.1
pytube==15.0.0
reportlab==4.3.0
SQLAlchemy==2.0.36
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
weasyprint==64.0
webencodings==0.5.1
Werkzeug==3.1.3
youtube-dl==2021.12.17
yt-dlp==2024.12.23
zopfli==0.2.3.post1
